more more more
git-svn-id: https://svn.code.sf.net/p/linux-gpib/code/trunk@13 b61deed1-901b-0410-a6dd-e94ecbdaea89

+++ linux-gpib/driver/sys/ibsys.h
@@ -48,7 +48,7 @@
 
 
 
-extern uint16      ibbase;	/* base addr of GPIB interface registers  */
+extern unsigned int      ibbase;	/* base addr of GPIB interface registers  */
 extern uint8       ibirq;	/* interrupt request line for GPIB (1-7)  */
 extern uint8       ibdma ;      /* DMA channel                            */
